Commented out some errors to address later
