+++ CameraViewer/CameraViewer/Samples/CameraControlTool/Camera Viewer.csproj
@@ -95,6 +95,7 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="SaveLoad.cs" />
     <EmbeddedResource Include="Engine Parts.resx">
       <DependentUpon>Engine Parts.cs</DependentUpon>
     </EmbeddedResource>
